refactor: remove redundant request and response types
