+++ wallet/src/program_facades/token.rs
@@ -1,4 +1,4 @@
-use common::{error::ExecutionFailureKind, sequencer_client::json::SendTxResponse};
+use common::{error::ExecutionFailureKind, rpc_primitives::requests::SendTxResponse};
 use nssa::{AccountId, program::Program};
 use nssa_core::{
     NullifierPublicKey, SharedSecretKey, encryption::IncomingViewingPublicKey,
